aidid_house/requirements.txt
scrapy
mysql-connector-python
configparser
scrapeops-scrapy
psycopg2